+++ src/kernels.js
@@ -19,13 +19,46 @@ import {
 import * as web3Helpers from './helpers/web3';
 
 /**
+ * Get kernels count from PandoraMarket contract
+ * 
+ * @param {Object} config Library config (provided by the proxy but can be overridden)
+ * @returns {Promise} A Promise object represents the {number} 
+ */
+export const fetchCount = async (config = {}) => {
+
+    expect.all(config, {
+        'web3': {
+            type: 'object',
+            code: WEB3_REQUIRED
+        },
+        'contracts.PandoraMarket.abi': {
+            type: 'object',
+            code: CONTRACT_REQUIRED,
+            args: ['PandoraMarket']
+        },
+        'addresses.PandoraMarket': {
+            type: 'string',
+            code: ADDRESS_REQUIRED,
+            args: ['PandoraMarket']
+        }
+    });
+
+    const mar = new config.web3.eth.Contract(config.contracts.PandoraMarket.abi, config.addresses.PandoraMarket);
+    const count = await mar.methods
+        .kernelsCount()
+        .call();
+
+    return Number.parseInt(count, 10);
+};
+
+/**
  * Get Kernel address by kernel id
  * 
  * @param {number} id
  * @param {Object} config Library config (provided by the proxy but can be overridden)
  * @returns {Promise} A Promise object represents the {string}
  */
-export const fetchAddressById = async (id, config) => {
+export const fetchAddressById = async (id, config = {}) => {
 
     expect.all(config, {
         'web3': {
@@ -192,6 +225,24 @@ export const fetchKernel = async (address = '', config = {}) => {
 };
 
 /**
+ * Get kernel by id
+ * 
+ * @param {integer} id 
+ * @param {Object} config Library config (provided by the proxy but can be overridden)
+ * @returns {Promise} A Promise object represents the {Object}
+ */
+export const fetchKernelById = async (id, config = {}) => {
+    
+    const address = await fetchAddressById(id, config);
+    const kernel = await fetchKernel(address, config);
+
+    return {
+        id: id,
+        ...kernel
+    };
+};
+
+/**
  * Get all kernels
  * 
  * @param {Object} config Library config (provided by the proxy but can be overridden)
@@ -199,42 +250,35 @@ export const fetchKernel = async (address = '', config = {}) => {
  */
 export const fetchAll = async (config = {}) => {
 
-    let id = 0;
     let records = [];
     let error = [];
 
     try {
 
-        // @todo Add method getKernelsCount to the PandoraMarket contract for avoid iterating with "try catch"
-        while (true) {
-            
-            const kernelAddress = await fetchAddressById(id++, config);// can be 0x0
-            
-            if (+kernelAddress === 0) {
-                break;
-            }
+        const count = await fetchCount(config);
 
+        for (let i=0; i < count; i++) {
+            
             try {
 
-                const kernelObj = await fetchKernel(kernelAddress, config);
+                const kernel = await fetchKernelById(i, config);
 
                 records.push({
-                    id: id,
-                    ...kernelObj
+                    id: i,
+                    ...kernel
                 });
             } catch(err) {
-                
                 error.push({
-                    address: kernelAddress,
+                    id: i,
                     message: err.message
                 });
-            }
+            }        
         }
     } catch(err) {
         error.push({
             error: err.message
         });
-    }
+    }   
 
     return {
         records,
@@ -322,13 +366,47 @@ export const addToMarket = (kernelContractAddress, publisherAddress, config = {}
 });
 
 /**
+ * Remove kernel from PandoraMarket
+ * 
+ * @param {String} kernelAddress
+ * @param {String} publisherAddress 
+ * @param {Object} config Library config (provided by the proxy but can be overridden) 
+ */
+export const removeKernel = (kernelAddress, publisherAddress, config = {}) => new Promise((resolve, reject) => {
+
+    expect.all(config, {
+        'web3': {
+            type: 'object',
+            code: WEB3_REQUIRED
+        },
+        'contracts.PandoraMarket.abi': {
+            type: 'object',
+            code: CONTRACT_REQUIRED,
+            args: ['PandoraMarket']
+        },
+        'addresses.PandoraMarket': {
+            type: 'string',
+            code: ADDRESS_REQUIRED,
+            args: ['Kernel']
+        }
+    });
+
+    const market = new config.web3.eth.Contract(config.contracts.PandoraMarket.abi, config.addresses.PandoraMarket);
+    market.methods
+        .removeKernel(kernelAddress)
+        .send({
+            from: publisherAddress
+        })
+        .on('error', reject)
+        .on('receipt', resolve);
+});
+
+/**
  * Handle event KernelAdded
  * 
- * @param {Function} storeCallback 
- * @param {Function} errorCallback
  * @param {Object} config Library config (provided by the proxy but can be overridden)
  */
-export const eventKernelAdded = (storeCallback = () => {}, errorCallback = () => {}, config = {}) => {
+export const eventKernelAdded = (config = {}) => new Promise((resolve, reject) => {
 
     expect.all(config, {
         'web3': {
@@ -354,15 +432,52 @@ export const eventKernelAdded = (storeCallback = () => {}, errorCallback = () =>
             try {
 
                 const kernel = await fetchKernel(res.returnValues.kernel, config);
-                storeCallback({
+                resolve({
                     address: res.returnValues.kernel,
                     kernel,
                     status: 'created',
                     event: 'PandoraMarket.KernelAdded'
                 });
             } catch(err) {
-                errorCallback(err);
+                reject(err);
             }            
         })
-        .on('error', errorCallback);
-};
+        .on('error', reject);
+});
+
+/**
+ * Handle event KernelRemoved
+ * 
+ * @param {Object} config Library config (provided by the proxy but can be overridden)
+ */
+export const eventKernelRemoved = (config = {}) => new Promise((resolve, reject) => {
+
+    expect.all(config, {
+        'web3': {
+            type: 'object',
+            code: WEB3_REQUIRED
+        },
+        'contracts.PandoraMarket.abi': {
+            type: 'object',
+            code: CONTRACT_REQUIRED,
+            args: ['PandoraMarket']
+        },
+        'addresses.PandoraMarket': {
+            type: 'string',
+            code: ADDRESS_REQUIRED,
+            args: ['Kernel']
+        }
+    });
+
+    const mar = new config.web3.eth.Contract(config.contracts.PandoraMarket.abi, config.addresses.PandoraMarket);
+    mar.events.KernelRemoved()
+        .on('data', async res => {
+
+            resolve({
+                address: res.returnValues.kernel,
+                status: 'removed',
+                event: 'PandoraMarket.KernelRemoved'
+            });            
+        })
+        .on('error', reject);
+});